Made sizzle throw when you specify an index and it's not found
The problem was that when you specified an index
you would be expecting an actual DOM node to be
returned, but if the index was not found it would
return a jquery object instead of a DOM node.
This would cause code to think a DOM node was
returned and try to access DOM properties that
didn't actually exist.